refactor: extract deal/restaurant extraction into own file
